Update france_jupyter_matplotlib.ipynb
I broke down the coding into mini chunks, supported with Markdown text. Also, renamed the variables,  for a less crowded code.
I could not read one of the files into  the code...any help would be appreciated, ofc please do not hesitate to highlight  and bring it up to the light.